fix some tests as RandomAccessFuseFileStream supports truncate and append

+++ dora/integration/fuse/src/main/java/alluxio/fuse/file/RandomAccessFuseFileStream.java
@@ -17,6 +17,8 @@ import alluxio.client.file.FileInStream;
 import alluxio.client.file.FileOutStream;
 import alluxio.client.file.FileSystem;
 import alluxio.exception.AlluxioException;
+import alluxio.exception.runtime.UnimplementedRuntimeException;
+import alluxio.fuse.AlluxioFuseUtils;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -27,6 +29,7 @@ import java.io.IOException;
 import java.io.RandomAccessFile;
 import java.nio.ByteBuffer;
 import java.nio.channels.FileChannel;
+import java.util.Optional;
 
 /**
  * FUSE file stream for supporting random access.
@@ -165,6 +168,37 @@ public class RandomAccessFuseFileStream implements FuseFileStream {
   public void truncate(long size) {
     initTmpFileIfNotInitialized();
 
+    /*
+    if (size == 0) {
+      try {
+        mRandomAccessFile.close();
+      } catch (IOException e) {
+
+      } finally {
+        mTmpFile.delete();
+      }
+
+      AlluxioFuseUtils.deletePath(mFileSystem, mURI);
+
+      try (FileOutStream fos = mFileSystem.createFile(mURI)) {
+        mHasInitializedTmpFile = false;
+      } catch (IOException | AlluxioException e) {
+      }
+      return;
+    }
+
+    try {
+      if (size > mRandomAccessFile.length()) {
+        throw new UnimplementedRuntimeException(
+            String.format("Cannot truncate file %s from size %s to size %s", mURI, size,
+                size));
+      }
+    } catch (IOException e) {
+      LOG.error("Failed to get mRandomAccessFile.length(). The temporary file path is {}. ",
+          mTmpFile, e);
+    }
+    */
+
     try {
       mRandomAccessFile.setLength(size);
     } catch (Exception e) {
@@ -176,27 +210,31 @@ public class RandomAccessFuseFileStream implements FuseFileStream {
 
   @Override
   public void close() {
-    // TODO(JiamingMai): rename it to a tmp file before deleting the file
-    try {
-      mFileSystem.delete(mURI);
-    } catch (IOException | AlluxioException e) {
-      LOG.error("Failed to delete file {} ", mURI);
+    if (mRandomAccessFile != null) {
+      // TODO(JiamingMai): rename it to a tmp file before deleting the file
+      try {
+        mFileSystem.delete(mURI);
+      } catch (IOException | AlluxioException e) {
+        LOG.error("Failed to delete file {} ", mURI);
+      }
+
+      // write the contents of the temporary file back to the file
+      try (FileOutStream fos = mFileSystem.createFile(mURI)) {
+        mRandomAccessFile.seek(0);
+        FileChannel channel = mRandomAccessFile.getChannel();
+        ByteBuffer buf = ByteBuffer.allocate(COPY_TO_LOCAL_BUFFER_SIZE_DEFAULT);
+        while (channel.read(buf) != -1) {
+          buf.flip();
+          fos.write(buf.array(), 0, buf.limit());
+        }
+        mRandomAccessFile.close();
+      } catch (IOException | AlluxioException e) {
+        LOG.error("Encountered exception when closing the stream:\n {} ", e.getMessage(), e);
+        throw new RuntimeException(e);
+      }
     }
 
-    // write the contents of the temporary file back to the file
-    try (FileOutStream fos = mFileSystem.createFile(mURI)) {
-      mRandomAccessFile.seek(0);
-      FileChannel channel = mRandomAccessFile.getChannel();
-      ByteBuffer buf = ByteBuffer.allocate(COPY_TO_LOCAL_BUFFER_SIZE_DEFAULT);
-      while (channel.read(buf) != -1) {
-        buf.flip();
-        fos.write(buf.array(), 0, buf.limit());
-      }
-      mRandomAccessFile.close();
-    } catch (IOException | AlluxioException e) {
-      LOG.error("Encountered exception when closing the stream:\n {} ", e.getMessage(), e);
-      throw new RuntimeException(e);
-    } finally {
+    if (mTmpFile != null) {
       mTmpFile.delete();
     }
   }
